Move lookup API mocks to their own file
